Move DataLocale to icu_locale_core
