+++ components/locale_core/src/lib.rs
@@ -70,12 +70,14 @@ extern crate alloc;
 #[macro_use]
 mod helpers;
 
+mod data;
 mod langid;
 mod locale;
 mod macros;
 mod parser;
 mod shortvec;
 
+pub use data::DataLocale;
 pub use langid::LanguageIdentifier;
 pub use locale::Locale;
 pub use parser::errors::ParseError;
